Change worker list sort: coord before workers
Before this change, we sort the list of worker nodes only by their host
name and port number. This was good enough when we only had worker nodes
in the metadata in the past, but now that we also have coordinator it is
harder to prevent distributed deadlocks. Hence, we now make sure that
the coordinator node(s) is always sorted to be in the beginning of the
list of worker nodes.

A DDL command that is run at the coordinator node will take locks on the
coordinator, and if successful attempt to take locks on the worker
nodes. Similarly, a DDL command that is run on a worker node will take
the locks on itself first, and then attempt to take locks on the
remaining of the workers. If the worker nodes are sorted before the
coordinator node, then we can end up in a situation where the
coordinator node is waiting for a lock on a worker node, while the
worker node is waiting for a lock on the coordinator node. This is a
distributed deadlock.