Fix isolation_add_remove_node test
We shouldn't be able to disable the first node in the cluster.
Isolation test doesn't handle cache invalidations/concurrency
very well, they are NOT real client backends.
It is possible that they improved isolation tester with PG16.

So, we fix the test here such that it doesn't disable the first node.